+++ libraries/template-haskell/changelog.md
@@ -1,5 +1,11 @@
 # Changelog for [`template-haskell` package](http://hackage.haskell.org/package/template-haskell)
 
+## 2.13.0.0 *TBA*
+
+  * Bundled with GHC *TBA*
+
+  * Add support for overloaded labels. Introduces `labelE :: String -> ExpQ`.
+
 ## 2.12.0.0 *TBA*
 
   * Bundled with GHC *TBA*